+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - OTLP logs being rejected or ingested with an incorrect timestamp if they only contain an observed time. ([#6290](https://github.com/getsentry/relay/pull/6290))
 - Fix streamed attachments having wrong retention. ([#6268](https://github.com/getsentry/relay/pull/6268))
 - Re-parameterize a DSC when it is computed from a transaction. ([#6279](https://github.com/getsentry/relay/pull/6279))
+- Correct error and outcome for too large streams. ([#6291](https://github.com/getsentry/relay/pull/6291))
 - Validate streamed minidumps and correctly enforce size limit for buffered minidumps. ([#6282](https://github.com/getsentry/relay/pull/6282))
 
 ## 26.7.2